+++ apps/fluent-tester/src/TestComponents/MenuButtonV1/testData.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-import { MenuButtonItemProps } from '@fluentui/react-native';
+import type { MenuButtonItemProps } from '@fluentui/react-native';
 import { testImage, svgProps } from '../Common/iconExamples';
 
 export const menuItems: MenuButtonItemProps[] = [