+++ apps/fluent-tester/src/TestComponents/Svg/SvgTest.tsx
@@ -4,7 +4,8 @@ import { Separator } from '@fluentui/react-native';
 import { Circle, Defs, G, Line, Path, Polygon, LinearGradient, RadialGradient, Rect, Stop, Svg, SvgUri, Use } from 'react-native-svg';
 import TestSvg from './Assets/accessible-icon-brands.svg';
 import { SVG_TESTPAGE } from '../../../../E2E/src/Svg/consts';
-import { Test, TestSection, PlatformStatus } from '../Test';
+import type { TestSection, PlatformStatus } from '../Test';
+import { Test } from '../Test';
 const styles = StyleSheet.create({
   svg: {
     backgroundColor: 'green',