+++ apps/fluent-tester/src/theme/applyTheme.ts
@@ -1,4 +1,4 @@
-import { PartialTheme, Theme, ThemeOptions } from '@fluentui-react-native/framework';
+import type { PartialTheme, Theme, ThemeOptions } from '@fluentui-react-native/framework';
 import { createOfficeTheme, getThemingModule } from '@fluentui-react-native/win32-theme';
 
 export type ThemeNames = 'Default' | 'Office' | 'Caterpillar' | 'Apple';