+++ packages/deprecated/foundation-tokens/src/index.ts
@@ -1,4 +1,15 @@
-export * from './Token.types';
-export * from './Token';
-export * from './Token.function';
+export type {
+  IComponentTokens,
+  ILookupThemePart,
+  IOperationSet,
+  IStyleFactories,
+  IStyleFactoryEntry,
+  IStyleFactoryFunction,
+  IStyleFactoryFunctionRaw,
+  IStyleFactoryOperation,
+  IStyleFinalizer,
+  ITargetHasToken,
+} from './Token.types';
+export { processTokens } from './Token';
+export { buildComponentTokens } from './Token.function';
 export { styleFunction } from '@fluentui-react-native/tokens';