+++ packages/experimental/ActivityIndicator/src/ActivityIndicator.types.tsx
@@ -1,5 +1,5 @@
-import { Animated, ActivityIndicatorProps as CoreActivityIndicatorProps } from 'react-native';
-import { SvgProps } from 'react-native-svg';
+import type { Animated, ActivityIndicatorProps as CoreActivityIndicatorProps } from 'react-native';
+import type { SvgProps } from 'react-native-svg';
 
 export const activityIndicatorName = 'ActivityIndicator';
 /**