+++ packages/experimental/Expander/src/index.ts
@@ -1,2 +1,3 @@
-export * from './Expander';
-export * from './Expander.types';
+export { Expander } from './Expander';
+export { expanderName } from './Expander.types';
+export type { ExpandDirection, ExpanderProps, ExpanderTokens, ExpanderType, ExpanderViewProps, HorizontalAlignment, VerticalAlignment } from './Expander.types';