+++ packages/experimental/Shadow/src/Shadow.types.ts
@@ -1,5 +1,5 @@
-import { ShadowToken } from '@fluentui-react-native/theme-types';
-import { ViewProps } from 'react-native';
+import type { ShadowToken } from '@fluentui-react-native/theme-types';
+import type { ViewProps } from 'react-native';
 
 export const shadowName = 'Shadow';