+++ packages/experimental/Stack/src/Stack.tsx
@@ -1,10 +1,13 @@
 /** @jsx withSlots */
 import * as React from 'react';
 import StackItem from './StackItem/StackItem';
-import { StackProps, stackName, StackType, StackTokens } from './Stack.types';
-import { View, ViewProps } from 'react-native';
+import type { StackProps, StackType, StackTokens } from './Stack.types';
+import { stackName } from './Stack.types';
+import type { ViewProps } from 'react-native';
+import { View } from 'react-native';
 import { filterViewProps } from '@fluentui-react-native/adapters';
-import { compose, withSlots, mergeProps, UseSlots, getMemoCache } from '@fluentui-react-native/framework';
+import type { UseSlots } from '@fluentui-react-native/framework';
+import { compose, withSlots, mergeProps, getMemoCache } from '@fluentui-react-native/framework';
 import { stylingSettings } from './Stack.styling';
 
 // Needed for TS to understand that __jsiExecutorDescription exists.