+++ packages/theming/default-theme/src/createDefaultTheme.ts
@@ -1,7 +1,7 @@
 import { ThemeReference } from '@fluentui-react-native/theme';
 import { defaultFluentDarkTheme, defaultFluentHighConstrastTheme, defaultFluentTheme } from './defaultTheme';
 import { Appearance } from 'react-native';
-import { Theme, ThemeOptions } from '@fluentui-react-native/theme-types';
+import type { Theme, ThemeOptions } from '@fluentui-react-native/theme-types';
 import { getCurrentAppearance } from '@fluentui-react-native/theming-utils';
 import assertNever from 'assert-never';