+++ packages/experimental/Shimmer/CHANGELOG.md
@@ -1,8 +1,29 @@
 # Change Log - @fluentui-react-native/experimental-shimmer
 
-This log was last generated on Mon, 30 Jan 2023 07:12:38 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Feb 2023 01:29:39 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.16
+
+Thu, 02 Feb 2023 01:29:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.39
+- Bump @fluentui-react-native/use-styling to v0.9.3
+
+## 0.8.15
+
+Wed, 01 Feb 2023 22:33:17 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.38
+- Bump @fluentui-react-native/tokens to v0.20.7
+- Bump @fluentui-react-native/theming-utils to v0.23.2
+- Bump @fluentui-react-native/component-cache to v1.4.3
+- Bump @fluentui-react-native/use-styling to v0.9.2
 
 ## 0.8.14